% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convGLM_vcov.R
\name{convGLM_vcov}
\alias{convGLM_vcov}
\title{Variance covariance matrix for propensity-weighted generalized linear models with logistic weight estimation}
\usage{
convGLM_vcov(estwt_fit, fit_outcome, biased)
}
\arguments{
\item{estwt_fit}{A \code{glm} object from the propensity weight estimation}

\item{fit_outcome}{A \code{glm} object from the final outcome model}

\item{biased}{An indicator of biased sample membership}
}
\value{
Covariance matrix estimate that accounts for uncertainty in propensity-weight estimation
for model coefficients from \code{fit_outcome}
}
\description{
Estimate covariance matrix with a simultaneous estimating equation approach for a propensity-weighted generalized linear model
when weights are estimated via logistic regression. Note that this is the same as logisticSE() but
it returns the covariance matrix instead of standard errors.
}
