% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticSE.R
\name{logisticSE}
\alias{logisticSE}
\title{Standard error estimation for propensity-weighted generalized linear models with logistic weight estimation}
\usage{
logisticSE(estwt_fit, fit_outcome, biased)
}
\arguments{
\item{estwt_fit}{A \code{glm} object from the propensity weight estimation}

\item{fit_outcome}{A \code{glm} object from the final outcome model}

\item{biased}{An indicator of biased sample membership}
}
\value{
Standard error estimates that account for uncertainty in propensity-weight estimation
for model coefficients from \code{fit_outcome}
}
\description{
Estimate standard errors with a simultaneous estimating equation approach for a propensity-weighted generalized linear model
when weights are estimated via logistic regression
}
