% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentSE.R
\name{treatmentSE}
\alias{treatmentSE}
\title{Standard error estimation for treatment effects estimated with convenience samples}
\usage{
treatmentSE(estwt_fit, estprop_fit, fit_outcome, biased, outcome_family)
}
\arguments{
\item{estwt_fit}{A \code{glm} object from the sampling weight estimation}

\item{estprop_fit}{A \code{svyglm} object from the propensity score estimation}

\item{fit_outcome}{A \code{svyglm} object from the treatment effect estimation}

\item{biased}{An indicator of biased sample membership}

\item{outcome_family}{Error distribution and link function to be used in the outcome model
(See \code{family} for details of family function). Defaults to stats::gaussian}
}
\value{
Standard error estimates for treatment effect estimates that account for
uncertainty in sampling-weight estimation, propensity score estimation, and
estimating the treatment effect
}
\description{
Estimate standard errors with a simultaneous estimating equation approach for
propensity adjusted estimates of a treatment effect when sampling weights are estimated
using an auxiliary sample and the sampling weights are used to estimate the propensity score
and the treatment effect
}
